requirements.txt
colorama==0.4.4
grpcio==1.37.1
grpcio-tools==1.37.1
loguru==0.5.3
passlib==1.7.4
peewee==3.14.4
protobuf==3.15.8
PyMySQL==1.0.2
six==1.15.0
win32-setctime==1.0.3
